fix mac compile
Former-commit-id: a38b3e35f08cc2f409d709c10e1c44467613ab08

+++ src/cert.cpp
@@ -1248,7 +1248,7 @@ UniValue certcount(const UniValue& params, bool fHelp) {
 			throw runtime_error("SYSCOIN_CERTIFICATE_RPC_ERROR: ERRCODE: 2517 - " + _("Scan failed"));
 	}
 
-    return certScan.size();
+    return (int)certScan.size();
 }
 UniValue certlist(const UniValue& params, bool fHelp) {
 	if (fHelp || 4 < params.size())
